fix: Covering multiple serialization cases (#245)
* fix/serialization: missing serializer impl

* feat/octree: add clear

* refactor/scene: observer adder

* refactor/scene: remove light register helper func

* fix/scene: double registration of lights, hide synchronization

* feat/sceneManager: runtime add/set scene

* fix/scene: cover multiple branching (sync, add) case

* chore/sceneManager: fix comment

* refactor/client/scene: implement virtual func

* chore/client/scene: renamed func

* chore/engine: type casting warnings

* fix/sceneManager: load popup does not showed up

* bug/object: component deserialized before owner initialization

* fix/client/darkscene: finding nonexist skybox material

+++ Engine/Renderer.cpp
@@ -154,7 +154,7 @@ namespace Engine::Manager::Graphics
   ) const
   {
     StructuredBuffer<SBs::InstanceSB> sb;
-    sb.Create(structured_buffers.size(), structured_buffers.data(), false);
+    sb.Create(static_cast<UINT>(structured_buffers.size()), structured_buffers.data(), false);
     sb.BindSRV(SHADER_VERTEX);
 
     material->SetTempParam
