+++ Engine/egMesh.cpp
@@ -12,7 +12,7 @@ SERIALIZER_ACCESS_IMPL
 
 namespace Engine::Resources
 {
-  UINT Mesh::GetIndexCount() const { return m_indices_.size(); }
+  UINT Mesh::GetIndexCount() const { return static_cast<UINT>(m_indices_.size()); }
 
   const VertexCollection& Mesh::GetVertexCollection() const { return m_vertices_; }
 
@@ -151,14 +151,14 @@ namespace Engine::Resources
     GetD3Device().CreateBuffer<VertexElement>
       (
        D3D11_BIND_VERTEX_BUFFER,
-       m_vertices_.size(),
+       static_cast<UINT>(m_vertices_.size()),
        m_vertex_buffer_.ReleaseAndGetAddressOf(),
        m_vertices_.data()
       );
 
     GetD3Device().CreateBuffer<UINT>
       (
-       D3D11_BIND_INDEX_BUFFER, m_indices_.size(),
+       D3D11_BIND_INDEX_BUFFER, static_cast<UINT>(m_indices_.size()),
        m_index_buffer_.ReleaseAndGetAddressOf(), m_indices_.data()
       );
   }
